pyproject.toml
[tool.poetry]
name = "flask-mongodb-boilerplate"
version = "0.1.0"
description = ""
authors = ["Jorge Escobar <jorge@fromzero.io>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "2.0.2"
python-dotenv = "0.19.2"
flask-mongoengine = "1.0.0"
pytest = "6.2.1"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
black = {version = "^21.7b0", allow-prereleases = true}

[tool.black]
line-length = 80

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"